requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==5.0.3
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.13.1
identify==2.5.35
inflection==0.5.1
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
